feat: complete 100% dynamic registry refactor for configuration, models, and templates
- Refactor SusiConfig, ModelInfo, and IntentBundles into DynamicRegistry (HashMap<String, Value>) for 0 hardcoded keys.\n- Migrate hardcoded chat formatting to dynamic string interpolation in ChatTemplateConfig.\n- Restore and fix lost File Rollback and Sandbox fallback functionalities.\n- Populate config.default.json with comprehensive inference endpoints (vLLM, Ollama, OpenAI, Anthropic, Gemini) and default chat templates.\n- Ensure 100% pass rate across all 37 test suites.

+++ src/gemi/hardware.rs
@@ -390,8 +390,8 @@ impl HardwareProfiler {
         let cfg = crate::sandbox::manager::SusiConfig::load_global().unwrap_or_default();
 
         let mut ladder = Vec::new();
-        for step in cfg.model_ladder {
-            if ram_gb >= step.min_ram_gb {
+        for step in cfg.model_ladder() {
+            if ram_gb as f32 >= step.min_ram_gb {
                 ladder.push(ModelLadderStep {
                     step: step.step,
                     label: step.label,
